sched/fair: fix misfit with PELT.
Commit 20017f33839a ("sched/fair: Only kick nohz balance when runqueue
has more than 1 task") disabled the nohz kick for LB when a rq has a
misfit task. The assumption is that this would be addressed in the
forced up-migration path. However, this path is WALT-specific, so
disabling the nohz kick breaks PELT.

Fix it by re-enabling the nohz_kick when there is a misfit task on the
rq.

Bug: 143472450
Test: 10/10 iterations of eas_small_to_big ended up up-migrating
Fixes: 20017f33839a ("sched/fair: Only kick nohz balance when runqueue
has more than 1 task")
Signed-off-by: Quentin Perret <qperret@google.com>
Signed-off-by: Vishalcj17 <vishalcj@aospa.co>
Change-Id: I321b673e77dd4236363cbb8ffe33d475a819cb64